+++ src/neptune/common/warnings.py
@@ -13,7 +13,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-__all__ = ["warn_once", "NeptuneDeprecationWarning", "NeptuneWarning"]
+__all__ = [
+    "warn_once",
+    "warn_about_unsupported_type",
+    "NeptuneDeprecationWarning",
+    "NeptuneWarning",
+    "NeptuneUnsupportedType",
+]
 
 import os
 import traceback
@@ -27,6 +33,10 @@ class NeptuneDeprecationWarning(DeprecationWarning):
 
 
 class NeptuneWarning(Warning):
+    pass
+
+
+class NeptuneUnsupportedType(Warning):
     pass
 
 
@@ -44,7 +54,10 @@ def get_user_code_stack_level():
     return 2
 
 
-def warn_once(message: str, *, exception: type(Exception) = NeptuneDeprecationWarning):
+def warn_once(message: str, *, exception: type(Exception) = None):
+    if exception is None:
+        exception = NeptuneDeprecationWarning
+
     if message not in warned_once:
         warnings.warn(
             message=message,
@@ -52,3 +65,13 @@ def warn_once(message: str, *, exception: type(Exception) = NeptuneDeprecationWa
             stacklevel=get_user_code_stack_level(),
         )
         warned_once.add(message)
+
+
+def warn_about_unsupported_type(type_str: str):
+    warn_once(
+        message=f"""You're attempting to log a type that is not directly supported by Neptune ({type_str}).
+        Convert the value to a supported type, such as a string or float, or use stringify_unsupported(obj)
+        for dictionaries or collections that contain unsupported values.
+        For more, see https://docs.neptune.ai/help/value_of_unsupported_type""",
+        exception=NeptuneUnsupportedType,
+    )