+++ CHANGELOG.md
@@ -18,10 +18,10 @@
 - Removed `Run._short_id` property ([#1234](https://github.com/neptune-ai/neptune-client/pull/1234))
 - Removed `get_run_url` method ([#1238](https://github.com/neptune-ai/neptune-client/pull/1238))
 - Removed `neptune.new.sync` module ([#1240](https://github.com/neptune-ai/neptune-client/pull/1240))
-- Added exception for unsupported types ([#1229](https://github.com/neptune-ai/neptune-client/pull/1229))
 - Change run status in the table returned by `fetch_runs_table` to Active / Inactive ([#1233](https://github.com/neptune-ai/neptune-client/pull/1233))
 - Package renamed from `neptune-client` to `neptune` ([#1225](https://github.com/neptune-ai/neptune-client/pull/1225))
 - Changed values used to filter runs table by state ([#1253](https://github.com/neptune-ai/neptune-client/pull/1253))
+- Added warning for unsupported types ([#1255](https://github.com/neptune-ai/neptune-client/pull/1255))
 
 ### Fixes
 - Fixed input value type verification for `append()` method ([#1254](https://github.com/neptune-ai/neptune-client/pull/1254))